Shows dropdown labels and enables autosearch in page with results
