+++ weblate/static/style-bootstrap.css
@@ -478,12 +478,16 @@ legend {
     width: 35px;
     padding-left: 0;
     padding-right: 0;
+    border-top-right-radius: 4px;
+    border-bottom-right-radius: 4px;
 }
 #query-sort-toggle span:not(.active) {
     display: none;
 }
 #query-sort-dropdown {
-    width: 115px;
+    width: 160px;
+    border-top-left-radius: 4px;
+    border-bottom-left-radius: 4px;
 }
 #is-exact span {
     margin: 4px 4px 0 4px;
