+++ weblate/templates/snippets/query-field.html
@@ -3,7 +3,7 @@
 {% load icons %}
 {% load crispy_forms_field %}
 <div class="row">
-  <div class="col-sm-10">
+  <div class="col-sm-9">
     <div class="form-group">
       <div class="search-group query-field" role="group" >
         <div class="input-group">
@@ -27,7 +27,7 @@
       </div>
     </div>
   </div>
-  <div class="col-sm-2">
+  <div class="col-sm-3">
     <div class="form-group">
       <div class="search-group sort-field" role="group">
         <div class="input-group">
@@ -43,6 +43,7 @@
             <input type="hidden" id="id_sort_by" name="sort_by" value="{{ sort_query|default:'position,-priority' }}" aria-label="{% trans "Sort By" %}" />
             <input type="hidden" id="id_sort_direction" name="sort_by_direction" value="{{ sort_direction|default:'' }}" aria-label="{% trans "Sort By" %}" />
             <ul class="dropdown-menu">
+              <li><a data-sort="position,-priority">{% trans "Position & Priority" %}</a></li>
               <li><a data-sort="position">{% trans "Position" %}</a></li>
               <li><a data-sort="priority">{% trans "Priority" %}</a></li>
               <li><a data-sort="num_words">{% trans "Word count" %}</a></li>
